+++ modules/testing/src/main/kotlin/app/epistola/suite/testing/TestcontainersConfiguration.kt
@@ -1,4 +1,4 @@
-package app.epistola.suite
+package app.epistola.suite.testing
 
 import org.springframework.boot.test.context.TestConfiguration
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection
@@ -8,16 +8,6 @@ import org.testcontainers.utility.DockerImageName
 
 @TestConfiguration(proxyBeanMethods = false)
 class TestcontainersConfiguration {
-
-// disabled as we currently do not need grafana in the tests
-//    @Bean
-//    @ServiceConnection
-//    fun grafanaLgtmContainer(): LgtmStackContainer = LgtmStackContainer(
-//        DockerImageName.parse("grafana/otel-lgtm:latest"),
-//    ).also {
-//        it.followOutput(Slf4jLogConsumer(LoggerFactory.getLogger("grafana.otel.lgtm")))
-//    }
-
     @Bean
     @ServiceConnection
     fun postgresContainer(): PostgreSQLContainer = PostgreSQLContainer(DockerImageName.parse("postgres:17"))